+++ src/pms.cpp
@@ -5,7 +5,12 @@
 #include "pms.h"
 int main() {
     std::cout << "------WELCOME TO THE PARKING MANAGEMENT SYSTEM------" << std::endl;
-    std::cout << "[CHOOSE YOUR MODE]\n*MODE 1*: DEMO MODE\n*MODE 2*: DECIDE YOUR PARKING LOT\n<<<enter your choice<<< ";
+    //decide simulation length
+    std::cout << "*NOTE* In this simulation [1 second] in REAL equals [10 minutes]\n[Decide on the total time of simulation](in seconds)>>>";
+    int sec = 0;
+    std::cin>>sec; std::cin.ignore(100,'\n');
+    //mode choose
+    std::cout << "[CHOOSE YOUR MODE]\n*MODE 1*: DEMO MODE\n*MODE 2*: DECIDE YOUR PARKING LOT\n[enter your choice]>>> ";
     int choice = 1;
     std::cin >> choice;
     //time seed
@@ -16,11 +21,15 @@ int main() {
     time_t start_time,cur_time;
     time(&start_time);
     time(&cur_time);
-    while(cur_time-start_time <= 60){
+    //main loop
+    while(cur_time-start_time <= sec){
         my_park.Generate_Vehicle();
         my_park.Checkout_Veh();
+        my_park.draw();
         delay(1);
         time(&cur_time);
     }
+    //end
+    std::cout << "End of simulation, bye~" << std::endl;
     return 0;
 }